requirement.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.22.2
scipy==1.7.3
matplotlib==3.5.1
python-dateutil==2.8.2
opencv-python==4.5.4.60
Flask==2.0.1
IPython==8.2.0

